% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_one.R
\name{add_one}
\alias{add_one}
\title{Add one or more pieces to the chessboard}
\usage{
add_one(data, name, x, y)
}
\arguments{
\item{data}{list}

\item{name}{character, name of the chess}

\item{x}{numeric, x-axis}

\item{y}{numeric, y-axis}
}
\value{
Add pieces to the data
}
\description{
Add one or more pieces to the chessboard
}
\examples{
initiate_chessboard() \%>\%
  add_one("MA", 5, 1)

pname <- c("BING", "ZU", "SHI", "SHI", "XIANG", "PAO")
px <- c(1,1,4,6,3,2)
py <- c(4,7,1,1,1,3)

initiate_chessboard() \%>\%
  add_one(pname, px, py)
}
