% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animate_chessboard.R
\name{animate_chessboard}
\alias{animate_chessboard}
\title{Animate the record}
\usage{
animate_chessboard(data, speed = 1, loop = FALSE)
}
\arguments{
\item{data}{list}

\item{speed}{numeric, control the speed of the animation}

\item{loop}{logical, control the loop of the animation}
}
\value{
Play the record
}
\description{
Animate the record
}
\examples{
initiate_chessboard() \%>\%
  add_one("JU", 1, 1) \%>\%
  add_one("MA", 2, 1) \%>\%
  add_one("PAO", 7, 1) \%>\%
  animate_chessboard()
}
