% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiate_chessboard.R
\name{initiate_chessboard}
\alias{initiate_chessboard}
\title{Demonstrate the chessboard with labeling points.}
\usage{
initiate_chessboard()
}
\value{
The chessboard
}
\description{
Demonstrate the chessboard with labeling points.
}
\examples{
initiate_chessboard()
}
