% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/player.R
\name{player}
\alias{player}
\title{Player input when race start}
\usage{
player(data, before, after)
}
\arguments{
\item{data}{list}

\item{before}{numeric vector of length 2}

\item{after}{numeric vector of length 2}
}
\value{
The record, pieces information, and the graph
}
\description{
Player input when race start
}
\examples{
race_start() \%>\%
player(c(2,1), c(3, 3)) \%>\%
player(c(8,8), c(5, 8)) \%>\%
player(c(2,3), c(2, 10))
}
