% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puzzle.R
\name{puzzle}
\alias{puzzle}
\title{Demonstrate the chessboard with labeling points.}
\usage{
puzzle()
}
\value{
The chessboard
}
\description{
Demonstrate the chessboard with labeling points.
}
\examples{
puzzle()
}
