% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replay.R
\name{replay}
\alias{replay}
\title{Replay the game}
\usage{
replay(data, speed = 1, loop = TRUE)
}
\arguments{
\item{data}{list}

\item{speed}{numeric, control the speed of the animation}

\item{loop}{logical, control the loop of the animation}
}
\value{
Replay the game
}
\description{
Replay the game
}
\examples{
race_start() \%>\%
  player(c(2,1), c(3, 3)) \%>\%
  player(c(8,8), c(5, 8)) \%>\%
  player(c(2,3), c(2, 10)) \%>\%
  replay()
}
