% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_current.R
\name{show_current}
\alias{show_current}
\title{Clean the data label of axes in current graph}
\usage{
show_current(data)
}
\arguments{
\item{data}{dataframe or list}
}
\value{
Show the records
}
\description{
Clean the data label of axes in current graph
}
\examples{
initiate_chessboard() \%>\%
  add_one("PAO", 5, 1) \%>\%
  add_one("JU", 6, 3) \%>\%
  show_current()

pname <- c("BING", "ZU", "SHI", "SHI", "XIANG", "PAO")
px <- c(1, 1, 4, 6, 3, 2)
py <- c(4, 7, 1, 1, 1, 3)

initiate_chessboard() \%>\%
  add_one(pname, px, py) \%>\%
  add_one("PAO", 5, 1) \%>\%
  show_current()
}
