% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_stream.R
\name{show_stream}
\alias{show_stream}
\title{Clean the data label of axes in each graph}
\usage{
show_stream(data)
}
\arguments{
\item{data}{dataframe or list}
}
\value{
Show the records
}
\description{
Clean the data label of axes in each graph
}
\examples{
initiate_chessboard() \%>\%
  add_one("PAO", 5, 1) \%>\%
  add_one("JU", 6, 3) \%>\%
  show_stream()

pname <- c("BING", "ZU", "SHI", "SHI", "XIANG", "PAO")
px <- c(1, 1, 4, 6, 3, 2)
py <- c(4, 7, 1, 1, 1, 3)

initiate_chessboard() \%>\%
  add_one(pname, px, py) \%>\%
  add_one("PAO", 5, 1) \%>\%
  show_stream()
}
