% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve.R
\name{solve}
\alias{solve}
\title{Solve puzzle}
\usage{
solve(.data)
}
\arguments{
\item{.data}{tibble or list}
}
\value{
The puzzle
}
\description{
Solve puzzle
}
\examples{
pname <- c("JIANG", "SHI", "SHI", "MA", "JIU", "che", "che", "TZU", "pau", "pau", "shuai")
px <- c( 5, 4, 5, 6, 5, 2, 5, 4, 5, 5, 4)
py <- c(10,10, 9,10, 8, 4, 2, 3, 3, 4, 1)

puzzle() \%>\%
  add_one(pname, px, py) \%>\%
  solve()

puzzle() \%>\%
  add_one_pro(pname, px, py) \%>\%
  solve()
}
